A few updates
The RetroPi case configuration is not working.... 
Starting to work on the devops images. As Raspberry Pi has an architecture of arm64, it is not compatible with a lot on images on Docker Hub.

We have to build manually the nexus/jenkins/sonarqueue images.

Nginx is avaliable in arm64/v6

The Jenkins image build. For now it install all the necessary software to run Jenkins on the image.